add a navbar with links & rename nomatch to notfound for more clarity on enduser experience
